Revert "Replacing rclone with wget."
This reverts commit a85d11667baa7f67faa93fc21cd18e3d50dff5cf.

wget does not provide a consistent sync solution for interrupted
downloads and can create corrupted files in some cases.

A http proxy is used instead, so syncing can continue using rclone.

"However, if the file is bigger on the server because it’s been changed,
as opposed to just appended to, you’ll end up with a garbled file. Wget
has no way of verifying that the local file is really a valid prefix of
the remote file. You need to be especially careful of this when using
‘-c’ in conjunction with ‘-r’, since every file will be considered as an
"incomplete download" candidate."
-- https://www.gnu.org/software/wget/manual/wget.html ('--continue').